Collapse and refactor util package items.
