+++ app/src/main/java/com/OxGames/Pluvia/ui/screen/settings/SettingsGroupDebug.kt
@@ -18,11 +18,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import coil.annotation.ExperimentalCoilApi
 import coil.imageLoader
-import com.OxGames.Pluvia.CrashHandler
 import com.OxGames.Pluvia.service.SteamService
 import com.OxGames.Pluvia.ui.component.dialog.CrashLogDialog
 import com.OxGames.Pluvia.ui.theme.settingsTileColors
 import com.OxGames.Pluvia.ui.theme.settingsTileColorsDebug
+import com.OxGames.Pluvia.utils.application.CrashHandler
 import com.alorma.compose.settings.ui.SettingsGroup
 import com.alorma.compose.settings.ui.SettingsMenuLink
 import com.winlator.core.FileUtils
