+++ app/src/main/java/com/OxGames/Pluvia/utils/SteamUtils.kt
@@ -3,6 +3,8 @@ package com.OxGames.Pluvia.utils
 import android.annotation.SuppressLint
 import android.content.Context
 import android.provider.Settings
+import android.text.Html
+import com.OxGames.Pluvia.Constants
 import com.OxGames.Pluvia.service.SteamService
 import `in`.dragonbra.javasteam.util.HardwareUtils
 import java.io.File
@@ -104,4 +106,22 @@ object SteamUtils {
 
         return androidId.hashCode()
     }
+
+    /**
+     * Gets the avatar URL from an avatar hash.
+     */
+    fun getAvatarURL(string: String?): String =
+        string.orEmpty()
+            .ifEmpty { null }
+            ?.takeIf { str -> str.isNotEmpty() && !str.all { it == '0' } }
+            ?.let { "${Constants.Persona.AVATAR_BASE_URL}${it.substring(0, 2)}/${it}_full.jpg" }
+            ?: Constants.Persona.MISSING_AVATAR_URL
+
+    fun fromHtml(string: String): String = Html.fromHtml(string, Html.FROM_HTML_MODE_LEGACY).toString()
+
+    /**
+     * Gets the profile URL from a steam id.
+     * Steam should redirect to a vanity URL if applied.
+     */
+    fun getProfileUrl(id: Long): String = "${Constants.Persona.PROFILE_URL}$id/"
 }
