--- app/src/main/java/com/OxGames/Pluvia/utils/StringUtils.kt
@@ -1,19 +1,0 @@
-package com.OxGames.Pluvia.utils
-
-import android.text.Html
-import com.OxGames.Pluvia.Constants
-
-/**
- * Extension functions relating to [String] as the receiver type.
- */
-
-fun String.getAvatarURL(): String =
-    this.ifEmpty { null }
-        ?.takeIf { str -> str.isNotEmpty() && !str.all { it == '0' } }
-        ?.let { "${Constants.Persona.AVATAR_BASE_URL}${it.substring(0, 2)}/${it}_full.jpg" }
-        ?: Constants.Persona.MISSING_AVATAR_URL
-
-fun String.fromHtml(): String = Html.fromHtml(this, Html.FROM_HTML_MODE_LEGACY).toString()
-
-// This doesn't belong here, but i'm tired.
-fun Long.getProfileUrl(): String = "${Constants.Persona.PROFILE_URL}$this/"
