+++ app/src/main/java/com/OxGames/Pluvia/utils/application/ReleaseTree.kt
@@ -1,4 +1,4 @@
-package com.OxGames.Pluvia
+package com.OxGames.Pluvia.utils.application
 
 import android.util.Log
 import timber.log.Timber
